refactor: split config.py into domain package and eliminate direct os.environ access
Converts src/pincer/config.py into a src/pincer/config/ package with five
domain classes (LLMSettings, ChannelSettings, ToolSettings, APISettings,
CoreSettings) that all feed into a single Settings class backed by
pydantic-settings. get_settings() and get_settings_relaxed() are re-exported
from config/__init__.py so all 44 existing import sites are unchanged.

Replaces direct os.environ.get() calls for PINCER_* vars in security/doctor.py,
api/server.py, mcp/sampling.py, integrations/ms365/config.py, and
integrations/slack/auth.py with the appropriate settings accessor. Variables
set only in .env (not exported to the shell) are now visible to all these
sites, fixing the false-negative reports from pincer doctor.

Closes #117
