+++ src/pincer/config/tools.py
@@ -1,0 +1,38 @@
+from __future__ import annotations
+
+from pydantic import BaseModel, Field, SecretStr
+
+
+class ToolSettings(BaseModel):
+    # ── Shell ─────────────────────────────────────────────
+    shell_enabled: bool = Field(default=True, description="Enable shell execution tool")
+    shell_timeout: int = Field(default=30, ge=5, le=300, description="Shell command timeout secs")
+    shell_require_approval: bool = Field(
+        default=True,
+        description="Require user approval before running shell commands",
+    )
+
+    # ── Search ────────────────────────────────────────────
+    tavily_api_key: SecretStr = Field(default=SecretStr(""), description="Tavily API key")
+
+    # ── Memory ────────────────────────────────────────────
+    memory_enabled: bool = Field(default=True, description="Enable memory system")
+    summary_model: str = Field(
+        default="claude-haiku-4-5-20251001",
+        description="Cheap model for conversation summarization",
+    )
+    summary_threshold: int = Field(default=20, ge=5, description="Summarize conversation after N messages")
+
+    # ── Image Generation ──────────────────────────────────
+    image_provider: str = Field(
+        default="auto",
+        description="Image provider: auto | fal | gemini",
+    )
+    fal_key: SecretStr = Field(default=SecretStr(""), description="fal.ai API key")
+    fal_model: str = Field(default="fal-ai/nano-banana-2", description="fal.ai image model")
+    gemini_api_key: SecretStr = Field(default=SecretStr(""), description="Google Gemini API key")
+    image_model_gemini: str = Field(default="gemini-2.5-flash-image", description="Gemini image generation model")
+    image_max_cost_per_request: float = Field(
+        default=0.10, ge=0.0, description="Max cost per image generation request in USD"
+    )
+    image_daily_limit: int = Field(default=50, ge=0, description="Max image generations per day (0 = unlimited)")
